Move some more mocks to the tests folder
