+++ listenbrainz/webserver/static/js/tests/__mocks__/userSocialNetworkProps.json
